+++ wrappers/go/foundation/ed25519.go
@@ -278,7 +278,7 @@ func (obj *Ed25519) ExportPrivateKey(privateKey PrivateKey) (*RawPrivateKey, err
 /*
 * Check if algorithm can encrypt data with a given key.
 */
-func (obj *Ed25519) CanEncrypt(publicKey PublicKey, dataLen uint32) bool {
+func (obj *Ed25519) CanEncrypt(publicKey PublicKey, dataLen int) bool {
     proxyResult := /*pr4*/C.vscf_ed25519_can_encrypt(obj.cCtx, (*C.vscf_impl_t)(unsafe.Pointer(publicKey.Ctx())), (C.size_t)(dataLen)/*pa10*/)
 
     runtime.KeepAlive(obj)
@@ -291,21 +291,21 @@ func (obj *Ed25519) CanEncrypt(publicKey PublicKey, dataLen uint32) bool {
 /*
 * Calculate required buffer length to hold the encrypted data.
 */
-func (obj *Ed25519) EncryptedLen(publicKey PublicKey, dataLen uint32) uint32 {
+func (obj *Ed25519) EncryptedLen(publicKey PublicKey, dataLen int) int {
     proxyResult := /*pr4*/C.vscf_ed25519_encrypted_len(obj.cCtx, (*C.vscf_impl_t)(unsafe.Pointer(publicKey.Ctx())), (C.size_t)(dataLen)/*pa10*/)
 
     runtime.KeepAlive(obj)
 
     runtime.KeepAlive(publicKey)
 
-    return uint32(proxyResult) /* r9 */
+    return int(proxyResult) /* r9 */
 }
 
 /*
 * Encrypt data with a given public key.
 */
 func (obj *Ed25519) Encrypt(publicKey PublicKey, data []byte) ([]byte, error) {
-    outBuf, outBufErr := bufferNewBuffer(int(obj.EncryptedLen(publicKey.(PublicKey), uint32(len(data))) /* lg2 */))
+    outBuf, outBufErr := bufferNewBuffer(int(obj.EncryptedLen(publicKey.(PublicKey), len(data)) /* lg2 */))
     if outBufErr != nil {
         return nil, outBufErr
     }
@@ -330,7 +330,7 @@ func (obj *Ed25519) Encrypt(publicKey PublicKey, data []byte) ([]byte, error) {
 * Check if algorithm can decrypt data with a given key.
 * However, success result of decryption is not guaranteed.
 */
-func (obj *Ed25519) CanDecrypt(privateKey PrivateKey, dataLen uint32) bool {
+func (obj *Ed25519) CanDecrypt(privateKey PrivateKey, dataLen int) bool {
     proxyResult := /*pr4*/C.vscf_ed25519_can_decrypt(obj.cCtx, (*C.vscf_impl_t)(unsafe.Pointer(privateKey.Ctx())), (C.size_t)(dataLen)/*pa10*/)
 
     runtime.KeepAlive(obj)
@@ -343,21 +343,21 @@ func (obj *Ed25519) CanDecrypt(privateKey PrivateKey, dataLen uint32) bool {
 /*
 * Calculate required buffer length to hold the decrypted data.
 */
-func (obj *Ed25519) DecryptedLen(privateKey PrivateKey, dataLen uint32) uint32 {
+func (obj *Ed25519) DecryptedLen(privateKey PrivateKey, dataLen int) int {
     proxyResult := /*pr4*/C.vscf_ed25519_decrypted_len(obj.cCtx, (*C.vscf_impl_t)(unsafe.Pointer(privateKey.Ctx())), (C.size_t)(dataLen)/*pa10*/)
 
     runtime.KeepAlive(obj)
 
     runtime.KeepAlive(privateKey)
 
-    return uint32(proxyResult) /* r9 */
+    return int(proxyResult) /* r9 */
 }
 
 /*
 * Decrypt given data.
 */
 func (obj *Ed25519) Decrypt(privateKey PrivateKey, data []byte) ([]byte, error) {
-    outBuf, outBufErr := bufferNewBuffer(int(obj.DecryptedLen(privateKey.(PrivateKey), uint32(len(data))) /* lg2 */))
+    outBuf, outBufErr := bufferNewBuffer(int(obj.DecryptedLen(privateKey.(PrivateKey), len(data)) /* lg2 */))
     if outBufErr != nil {
         return nil, outBufErr
     }
@@ -395,14 +395,14 @@ func (obj *Ed25519) CanSign(privateKey PrivateKey) bool {
 * Return length in bytes required to hold signature.
 * Return zero if a given private key can not produce signatures.
 */
-func (obj *Ed25519) SignatureLen(privateKey PrivateKey) uint32 {
+func (obj *Ed25519) SignatureLen(privateKey PrivateKey) int {
     proxyResult := /*pr4*/C.vscf_ed25519_signature_len(obj.cCtx, (*C.vscf_impl_t)(unsafe.Pointer(privateKey.Ctx())))
 
     runtime.KeepAlive(obj)
 
     runtime.KeepAlive(privateKey)
 
-    return uint32(proxyResult) /* r9 */
+    return int(proxyResult) /* r9 */
 }
 
 /*
@@ -491,40 +491,40 @@ func (obj *Ed25519) ComputeSharedKey(publicKey PublicKey, privateKey PrivateKey)
 * Return number of bytes required to hold shared key.
 * Expect Public Key or Private Key.
 */
-func (obj *Ed25519) SharedKeyLen(key Key) uint32 {
+func (obj *Ed25519) SharedKeyLen(key Key) int {
     proxyResult := /*pr4*/C.vscf_ed25519_shared_key_len(obj.cCtx, (*C.vscf_impl_t)(unsafe.Pointer(key.Ctx())))
 
     runtime.KeepAlive(obj)
 
     runtime.KeepAlive(key)
 
-    return uint32(proxyResult) /* r9 */
+    return int(proxyResult) /* r9 */
 }
 
 /*
 * Return length in bytes required to hold encapsulated shared key.
 */
-func (obj *Ed25519) KemSharedKeyLen(key Key) uint32 {
+func (obj *Ed25519) KemSharedKeyLen(key Key) int {
     proxyResult := /*pr4*/C.vscf_ed25519_kem_shared_key_len(obj.cCtx, (*C.vscf_impl_t)(unsafe.Pointer(key.Ctx())))
 
     runtime.KeepAlive(obj)
 
     runtime.KeepAlive(key)
 
-    return uint32(proxyResult) /* r9 */
+    return int(proxyResult) /* r9 */
 }
 
 /*
 * Return length in bytes required to hold encapsulated key.
 */
-func (obj *Ed25519) KemEncapsulatedKeyLen(publicKey PublicKey) uint32 {
+func (obj *Ed25519) KemEncapsulatedKeyLen(publicKey PublicKey) int {
     proxyResult := /*pr4*/C.vscf_ed25519_kem_encapsulated_key_len(obj.cCtx, (*C.vscf_impl_t)(unsafe.Pointer(publicKey.Ctx())))
 
     runtime.KeepAlive(obj)
 
     runtime.KeepAlive(publicKey)
 
-    return uint32(proxyResult) /* r9 */
+    return int(proxyResult) /* r9 */
 }
 
 /*
